Ensure all lua link modifier entries point to existing files (#851)
	* Reorganize javascripts tests a bit
